+++ modulo4/Projeto-to-do-list/.gitignore
@@ -1,3 +1,4 @@
 build
 node_modules
+package-lock.json
 env
